Move to warp verify. Move util fns to utils.ts
